+++ R/zzz.R
@@ -3,6 +3,7 @@ utils::globalVariables(
         "axis", "coordinate", "spline", "X", "Y", "fan_line",
         "radius", "angle", "CI_upper", "CI_lower",
         "CI_lower_X", "CI_lower_Y", "CI_upper_X", "CI_upper_Y",
-        "index", "X_new", "Y_new", "CI_X", "CI_Y", "se.fit"
+        "index", "X_new", "Y_new", "CI_X", "CI_Y", "se.fit",
+        ".index"
     )
 )